晚上日常提交1

+++ src/main/java/cn/hgxsp/miaosha_1/Domain/MiaoshaOrder.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class MiaoshaOrder {
 	private Long id;
-	private Long userId;
-	private Long  orderId;
-	private Long goodsId;
+	private int userId;
+	private int  orderId;
+	private int goodsId;
 
 }
